% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{central}
\alias{central}
\title{Indicateurs de tendance centrale}
\usage{
central(data, x, breaks)
}
\arguments{
\item{data}{un tibble}

\item{x}{la variable considérée (nécessairement numérique)}

\item{breaks}{un vecteur de limites de classes}
}
\value{
un tibble contenant trois variables, \code{name}, \code{x} et \code{y}
}
\description{
Calcule les trois indicateurs de tendance centrale (moyenne,
médiane et mode), ainsi que les densités correspondantes
}
\author{
Yves Croissant
}
