% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Chomage}
\alias{Chomage}
\title{Demandeurs d'emploi inscrits à Pôle emploi}
\format{
Un tibble de 2360 observations contenant :
\itemize{
\item date : le mois d'observation, de janvier 1996 à juillet 2020
\item sexe :  \code{homme}, \code{femme} ou \code{total}
\item age : la tranche d'âge, \verb{< 25}, \code{25-49}, \verb{>= 50} ou \code{total}
}
}
\source{
\href{https://statistiques.pole-emploi.org/stmt/static/methode_2018}{site de Pôle emploi}
}
\description{
Données mensuelles, par tranche d'âge et par sexe des demandeurs
d'emploi inscrits à Pôle emploi
}
