% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Communes}
\alias{Communes}
\title{Population et surface des communes françaises}
\format{
Un tibble contenant
\itemize{
\item code : le code postal de la commune
\item pop : la population en milliers d'habitants
\item surface : la surface en kilomètres carré
}
}
\source{
\href{https://www.insee.fr/fr/statistiques/2521169}{site de l'INSEE}
}
\description{
Population et surface des 34989 communes françaises
}
\keyword{datasets}
