% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cond_table}
\alias{cond_table}
\title{Statistique conditionnelle}
\usage{
cond_table(data, x, x1, x2 = NULL, fun = mean, na.rm = TRUE)
}
\arguments{
\item{data}{un tibble}

\item{x}{une variable numérique}

\item{x1}{une première variable catégorielle}

\item{x2}{une éventuelle seconde variable catégorielle}

\item{fun}{la fonction à appliquer (par défaut la moyenne)}

\item{na.rm}{la valeur par défaut est \code{TRUE}, les observations
pour lesquelles la valeur de \code{x} est manquante sont retirées de
l'échantillon}
}
\value{
un tibble
}
\description{
Calcule une statistique (par défaut la moyenne arithmétique) d'une
variable numérique pour chaque modalités d'une ou de deux variables
catégorielles
}
\examples{

cond_table(Salaires, heures, secteur)
cond_table(Salaires, heures, secteur, sexe)
cond_table(Salaires, heures, secteur, fun = var)

}
\author{
Yves Croissant
}
