% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Conscrits}
\alias{Conscrits}
\title{La taille des conscrits}
\format{
Un tibble contenant
\itemize{
\item taille: la taille du conscrits en centimètres
}
}
\source{
\href{https://uni-tuebingen.de/fakultaeten/wirtschafts-und-sozialwissenschaftliche-fakultaet/faecher/fachbereich-wirtschaftswissenschaft/wirtschaftswissenschaft/lehrstuehle/volkswirtschaftslehre/wirtschaftsgeschichte/forschung/data-hub-height/}{data hub: Height Universitat Tubingen}
}
\description{
Un échantillon de 21064 conscrits en bavarois au XIXeme siècle
}
\references{
\insertRef{BATE:MURR:00}{statecoaes}
}
\keyword{datasets}
