% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Consommation}
\alias{Consommation}
\title{Consommation en France}
\format{
Un tibble contenant :
\itemize{
\item an : l'année d'observation
\item c_FT : la consommation pour chaque fonction en prix courants,
\item p_FT : l'indice de prix à la consommation pour chaque fonction (base 100 en 2014
}

Les postes de consommation sont :
\itemize{
\item alimentation,
\item alcool et tabac,
\item habillement,
\item logement,
\item ammeublement,
\item santé,
\item transport,
\item communications,
\item loisirs,
\item éducation,
\item restauration,
\item services divers,
\item solde territorial
\item total
}
}
\source{
Consomation des ménages en 2019, comptes nationaux annuels - base 2014

\href{https://www.insee.fr/fr/statistiques/4494154?sommaire=4494218}{site de l'INSEE}
}
\description{
Données annuelles de la consommation des ménages en France,
nomenclature par fonctions (alimentation, habillement, etc.)
}
