% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cont_table}
\alias{cont_table}
\alias{mean.cont_table}
\alias{variance.cont_table}
\alias{stdev.cont_table}
\alias{covariance.cont_table}
\alias{joint}
\alias{marginal}
\alias{conditional}
\alias{pre_print.cont_table}
\alias{format.cont_table}
\title{Table de contingence}
\usage{
cont_table(
  data,
  y1,
  y2,
  pond = NULL,
  total = TRUE,
  first1 = NULL,
  last1 = NULL,
  inflate1 = NULL,
  first2 = NULL,
  last2 = NULL,
  inflate2 = NULL
)

\method{mean}{cont_table}(x, ...)

\method{mean}{cont_table}(x, ...)

\method{variance}{cont_table}(x, ...)

\method{stdev}{cont_table}(x, ...)

\method{covariance}{cont_table}(x, drop = TRUE, ...)

joint(x)

marginal(x, y = NULL)

conditional(x, y = NULL)

\method{pre_print}{cont_table}(x)

\method{format}{cont_table}(x, ..., n = NULL, width = NULL, n_extra = NULL)
}
\arguments{
\item{data}{un tibble}

\item{y1}{une première variable catégorielle}

\item{y2}{une seconde variable catégorielle}

\item{pond}{une éventuelle variable contenant les pondérations à
utiliser pour passer de l'échantillon à la population}

\item{total}{si \code{TRUE} (valeur par défaut), un total est ajouté au
tableau}

\item{first1}{the center of the first class for the first variable}

\item{last1}{the center of the last class for the first variable}

\item{inflate1}{the width of the last class for the first variable}

\item{first2}{the center of the first class for the second variable}

\item{last2}{the center of the last class for the second variable}

\item{inflate2}{the width of the last class for the second variable}

\item{x}{un tibble contenant la table de contingence}

\item{...}{d'autres arguments}

\item{drop}{if \code{TRUE}, the default, a numeric is returned,
otherwise the result is a tibble}

\item{y}{la variable sur laquelle on veut réaliser l'opération}

\item{n}{the number of lines to print}

\item{width}{the width of the table to print}

\item{n_extra}{extra n lines}
}
\value{
un tibble
}
\description{
Une table de contingence rassemble les fréquences de toutes les
combinaisons de deux variables catégorielles sous la forme d'un
tableau à double entrée, avec les différentes modalités de la
première (seconde) variable en ligne (colonne). Les fonctions
\code{joint}, \code{marginal} et \code{conditional} permettent de calculer ces
trois distributions à partir d'une table de contingence (en
indiquant la variable pour laquelle la distribution doit être
calculée pour les deux dernières). Une méthode \code{mean} est définie
pour calculer la moyenne.
}
\examples{

cont_table(Emploi, diplome, sexe)
cont_table(Emploi, diplome, sexe, pond = pondérations)
cont_table(Emploi, diplome, sexe) \%>\% conditional(sexe)
cont_table(Salaires, salaire, taille)
cont_table(Salaires, salaire, taille) \%>\% joint
cont_table(Salaires, salaire, taille) \%>\% joint \%>\% mean
cont_table(Salaires, salaire, taille) \%>\% marginal(taille)
cont_table(Salaires, salaire, taille) \%>\% conditional(taille) \%>\% mean

}
\author{
Yves Croissant
}
