% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_radar.R
\name{coord_radar}
\alias{coord_radar}
\title{Graphique de type radar}
\source{
\href{https://www.cmap.polytechnique.fr/~lepennec/en/post/radar/radarandparallelplots/}{page web de Erwan Le Pennec}
}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{voir \code{coord_polar}}

\item{start}{voir \code{coord_polar}}

\item{direction}{voir \code{coord_polar}}
}
\description{
Une variante de \code{coord_polar} qui remplace les lignes lisses en
segments de droite
}
