% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Deaton}
\alias{Deaton}
\title{Relation entre richesse et développement}
\format{
Un tibble de 179 observations contenant :
\itemize{
\item code : le code du pays
\item nom : le nom du pays
\item region : la région du monde dans laquelle est située le pays
\item incgp : le groupe de revenu du pays
\item gdp : le PIB par tête
\item pop : la population
\item le : l'espérance de vie
\item subnom : un sous-ensemble de 17 noms de pays
}
}
\source{
\href{https://www.https://databank.worldbank.org/home.aspx}{site de la banque mondiale}
}
\description{
Données sur l'espérance de vie et le PIB par tête, qui permettent
d'analyser la relation entre richesse et développement
}
