% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Ebay}
\alias{Ebay}
\title{Ventes de voitures d'occasion}
\format{
Un tibble contenant :
\itemize{
\item miles : le kilométrage,
\item bid : le montant de l'enchère gagnante,
\item year : l'année de mise en service de la voiture
}
}
\source{
Journal of Applied Econometrics \href{https://qed.econ.queensu.ca/jae/}{data archive}
}
\description{
Données sur des voitures d'occasion aux Etats-Unis vendues sur le site Ebay
}
\references{
\insertRef{LEWI:11}{statecoaes}
}
