% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FranceLT}
\alias{FranceLT}
\title{Données de long terme pour la France}
\format{
Un tibble contenant :
\itemize{
\item an : l'année d'observation,
\item loyer : indice des loyers parisiens, base 100 en 2000,
\item cpi : indice de prix à la consommation, base 100 en 2000,
\item pib : produit intérieur brut en valeur,
\item pop : population de la France,
\item or : prix de l'or, base 100 en 2000
\item tict : taux d'intérêt à court terme,
\item tilt : taux d'intérêt à long terme.
}
}
\source{
Ministère de la transition écologique et solidaire
\href{https://www.data.gouv.fr/fr/datasets/valeurs-immobilieres-economiques-et-financieres-de-1800-a-2015/}{site de data.gouv.fr}
}
\description{
Données macro-économiques annuelles pour la France de 1840 à 2015
}
