% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{freq_table}
\alias{freq_table}
\title{Tables de Fréquence}
\usage{
freq_table(
  data,
  x,
  abs = FALSE,
  pct = FALSE,
  cumul = FALSE,
  pond = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{data}{un tibble}

\item{x}{une variable catégorielle}

\item{abs}{une valeur logique : \code{TRUE} pour renvoyer des
fréquences absolues, \code{FALSE} (la valeur par défaut) pour des
fréquences absolues,}

\item{pct}{une valeur logique : \code{TRUE} pour renvoyer des
pourcentages (20.2 au lieu de 0.202)}

\item{cumul}{une valeur logique (\code{FALSE} est la valeur par
défaut). Si \code{TRUE}, les fréquences cummulées sont renvoyées en
plus des fréquences}

\item{pond}{une éventuelle variable contenant les pondérations à
utiliser pour passer de l'échantillon à la population}

\item{na.rm}{la valeur par défaut est \code{TRUE}, les observations
pour lesquelles la valeur de \code{x} est manquante sont retirées de
l'échantillon}
}
\value{
un tibble
}
\description{
Une table de fréquence est adaptée pour les variables
catégorielles, elle renvoit les différentes modalités et les
fréquences associées
}
\examples{

freq_table(Emploi, activité, abs = TRUE, cumul = FALSE)
freq_table(Emploi, activité, abs = TRUE, cumul = TRUE)
freq_table(Emploi, activité, abs = TRUE, cumul = FALSE, pond = pondérations)
freq_table(Emploi, activité, cumul = TRUE)
freq_table(Emploi, activité, cumul = TRUE, pond = pondérations)
}
\author{
Yves Croissant
}
