% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Gaz}
\alias{Gaz}
\title{Consommation horaire de gaz}
\format{
Un tibble contenant :
\itemize{
\item region : le code de la région
\item heure : le jour et l'heure
\item conso : la consommation
}
}
\source{
\href{https://www.data.gouv.fr/fr/datasets/consommation-journaliere-et-horaire-de-gaz-des-clients-grd-et-eld-par-region-reseaux-grtgaz-terega-donnees-definitives-1/}{site data.gouv.fr}
}
\description{
Consommation par heure de gaz pour les différentes régions française
}
