% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{variance}
\alias{stdev}
\alias{madev}
\alias{modval}
\alias{pre_print}
\alias{medial}
\title{Generic functions provided by the statecoaes package}
\usage{
variance(x, ..., na.rm = TRUE)

stdev(x, ..., na.rm = TRUE)

madev(x, ..., na.rm = TRUE)

modval(x, ...)

pre_print(x)

medial(x, tbl = FALSE)
}
\arguments{
\item{x}{the main argument, basically a tibble,}

\item{...}{further arguments}

\item{na.rm}{should the missing values be stripped before the
computation of the statistic,}

\item{tbl, }{if \code{FALSE} a numeric is returned, if \code{TRUE} a tibble
with the density is returned, which can be usefull for central
statistics like madev, modval and medial}
}
\value{
a numeric or a tibble
}
\description{
Some generics were created, some of them don't exist in base R (to
compute the mode and the mediale of a distribution for example) or
because the functions exist but are not generic (this is the case
of var, sd, mad, which are replaced by our variance, stdev and
madev generic functions
}
\author{
Yves Croissant
}
