% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hist2plot}
\alias{hist2plot}
\title{Put a tibble in form to plot}
\usage{
hist2plot(data, y = NULL, plot = "histogram")
}
\arguments{
\item{data}{a tibble returned by the \code{hist_table} function, it
should contains the center of the classes (\code{x}) and at least
one measure of the frequencies or densities (one of \code{f}, \code{n},
\code{p}, \code{d})}

\item{y}{mandatory argument if the tibble contains more than one
frequency or density}

\item{plot}{one of \code{histogram} (the default) and \code{freqpoly} ; in
the first case a tibble is returned with columns \code{x}, \code{y},
\code{xend}, \code{yend} and in the seconde case \code{x} and \code{y}.}
}
\value{
a tibble
}
\description{
Convert a tibble built using hist_table in a shape that make it
easy to plot
}
\details{
#'
}
\examples{
pad <- Padoue \%>\% hist_table(price, breaks = c(100, 200, 300, 400, 500, 1000), right = TRUE, cols = "Npd")
pad \%>\% hist2plot(y = "d") \%>\% ggplot() + geom_segment(aes(x, y, xend = xend, yend = yend))
pad \%>\% hist2plot(y = "d", plot = "freqpoly") \%>\% ggplot() + geom_line(aes(x, y))
}
\author{
Yves Croissant
}
