% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_table.R
\name{hist_table}
\alias{hist_table}
\title{Histogrammes}
\usage{
hist_table(
  data,
  x,
  cols = "n",
  vals = "x",
  breaks = NULL,
  xfirst = NULL,
  xlast = NULL,
  right = NULL,
  total = FALSE,
  inflate = NULL
)
}
\arguments{
\item{data}{un tibble}

\item{x}{la variable considérée (nécessairement numérique)}

\item{cols}{une chaîne de caractère contenant les lettres \code{n} pour
nombre, \code{f} pour fréquence et \code{p} pour pourcentage ; les séries
cumulées sont obtenues en indiquant les mêmes lettres en
majuscule. Par rapport à \code{freq_table}, la lettre \code{d} peut
également être indiquée pour calculer des densités}

\item{vals}{les valeurs de la variable renvoyées ; \code{x} pour le
centre de la classe, \code{l} et \code{u} pour les limites inférieure et
supérieure, \code{a} pour l'amplitude}

\item{breaks}{un vecteur de limites de classes}

\item{xfirst}{une valeur numérique indiquant le centre de la
première classe}

\item{xlast}{une valeur numérique indiquant le centre de la dernière
classe}

\item{right}{un booléen indiquant si les classes doivent être
fermées (\code{right = TRUE}) ou fermée (\code{right = FALSE}) à droite}

\item{total}{un total doit il être renvoyé ?}

\item{inflate}{dans le cas où la valeur centrale de la dernière
classe n'est pas renseignée, elle est fixée à la borne
inférieure plus ce coefficient multiplié par la moitié de
l'amplitude de la classe précédente}
}
\value{
un tibble contenant les valeurs de \code{vals} et de \code{cols}
spécifiées
}
\description{
Calcule les effectifs par classe pour une variable numérique
}
\examples{

# price is a numeric variable, a vector of breaks should be provided
Padoue \%>\% hist_table(price, breaks = c(50, 100, 150, 200, 250, 300, 350, 400), right = TRUE)
Padoue \%>\% hist_table(price, breaks = c(50, 100, 150, 200, 250, 300, 350, 400), right = TRUE, cols = "fd", vals = "xa")
# salaire is a factor that represents the classes
Salaires \%>\% hist_table(salaire, "d")
# a breaks argument is provided to reduce the number of classes
Salaires \%>\% hist_table(salaire, breaks = c(10, 20, 30, 40, 50))

}
\author{
Yves Croissant
}
