% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{indexes}
\alias{indexes}
\title{Indices synthétiques}
\usage{
indexes(data, an, bien, quant, prix, base, chaine = FALSE)
}
\arguments{
\item{data}{un tibble}

\item{an}{la série contenant les dates}

\item{bien}{la série contenant les biens}

\item{quant}{la série contenant les quantités}

\item{prix}{la série contenant les prix}

\item{base}{l'année de base de l'index}

\item{chaine}{si fixé à \code{TRUE}, un indice chaîné est calculé}
}
\value{
un tibble
}
\description{
Calcul des indices synthétiques (Laspeyres, Paasche and Fisher),
chaînés ou non
}
\author{
Yves Croissant
}
