% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{indices}
\alias{indices}
\title{Indices synthétiques}
\usage{
indices(data, an, bien, quant, prix, base, chaine = FALSE)
}
\arguments{
\item{data}{un tibble}

\item{an}{la date d'observation}

\item{bien}{le bien considéré}

\item{quant}{la quantité}

\item{prix}{le prix}

\item{base}{l'année de base}

\item{chaine}{si vrai, la formule de l'indice chaînée est utilisée}
}
\value{
un tibble
}
\description{
Calculs d'indices synthétiques (Laspeyres, Paasche et Fisher),
chaînés ou non.
}
\author{
Yves Croissant
}
