% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Logements}
\alias{Logements}
\title{Nombre de pièces des logements}
\format{
Un tibble contenant :
\itemize{
\item pieces : le nombre de pièces du logement,
\item nombre : le nombre de logements (en millions) pour chaque nombre
de pièces.
}
}
\source{
Logements et résidences principales en 2017,

Recensement de la population

site de l'\href{https://www.insee.fr/fr/statistiques/4515537?sommaire=4516107#consulter}{INSEE}.
}
\description{
Table de fréquence indiquant le nombre de pièce par logements
}
