% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Merlan}
\alias{Merlan}
\title{Pêche au Merlan}
\format{
Un tibble contenant :
\itemize{
\item debut : date de début de la semaine d'observation
\item fin : date de la fin de la semaine d'observation
\item poids : quantité de merlan pêchée en kg
\item valeur : valeur des ventes de merlan
\item id : identifiant du port
\item nom : nom du port
}
}
\source{
\href{https://www.data.gouv.fr/fr/datasets/serie-hebdomadaire-par-espece-produits-de-la-mer/}{site data.gouv.fr}
}
\description{
Pêche au Merlan (volume et valeur) par semaine pour les différents ports de pêche francais
}
