% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{moving_average}
\alias{moving_average}
\title{Moyennes mobiles}
\usage{
moving_average(data, freq = c("month", "quarter"))
}
\arguments{
\item{data}{un tibble}

\item{freq}{soit \code{month} pour des données mensuelles ou \code{quarter}
pour des données trimestrielles}
}
\value{
un tibble
}
\description{
Calcul des moyennes mobiles pour des données mensuelles ou
trimestrielles
}
\author{
Yves Croissant
}
