% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Naissance}
\alias{Naissance}
\title{Les naissances en 2017}
\format{
Un tibble contenant :
\itemize{
\item département : le code du département de naissance,
\item sexe : le sexe de l'enfant,
\item mois : le mois de la naissance,
\item age : l'âge de la mère.
}
}
\source{
Etat civil, site de l'\href{www.insee.fr}{INSEE}
}
\description{
Les fichiers de l'Etat-Civil recense de manière exhaustive les
naissances, les décès et les mariages.
}
