% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Naissances}
\alias{Naissances}
\title{Les naissances en 2017}
\format{
Un tibble contenant :
\itemize{
\item département : le code du département de naissance,
\item sexe : le sexe de l'enfant,
\item mois : le mois de la naissance,
\item age : l'âge de la mère,
\item tranche : tranche d'âge de la mère,
\item trimestre : trimestre de naissance
}
}
\source{
Naissances, décès et mariages en 2017,

Etat civil - Fichiers détail

site de l'\href{https://www.insee.fr/fr/statistiques/3596190?sommaire=3596198}{INSEE}
}
\description{
Les fichiers de l'Etat-Civil recense de manière exhaustive les
naissances, les décès et les mariages.
}
