% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Penalty}
\alias{Penalty}
\title{Séances de penalty}
\format{
Un tibble contenant
\itemize{
\item id : l'identifiant du match
\item order : \code{first} pour l'équipe qui débute la séance de tirs aux
buts, \code{second} pour l'autre équipe
\item rank : le rang du tir pour chaque équipe, 1 pour le premier
tireur, 2 pour le deuxième, etc.
\item cumrank : rang cumulé pour le tir au but, 1 pour le premier tir
de l'équipe qui commence la séance, 2 pour le premier tir de
l'autre équipe, 3 pour le second tir de l'équipe qui commence, etc.
\item score : 1 en cas de réussite du tir au but, 0 autrement.
}
}
\source{
\href{https://www.aeaweb.org/articles?id=10.1257/aer.100.5.2548}{site de l'AEA}
}
\description{
Résultat de chaque tir au but (réussite ou échec) dans le cadre de
séances de penalty permettant de départager deux équipes à égalité
}
\references{
\insertRef{APES:PALA:10}{statecoaes}
}
\keyword{datasets}
