% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Poissons}
\alias{Poissons}
\title{Pêche en France}
\format{
Un tibble contenant :
\itemize{
\item poisson : espèce de poisson
\item date : date de début de la semaine d'observation
\item poids : quantité de merlan pêchée en kg
\item valeur : valeur des ventes de merlan
}
}
\source{
\href{https://www.data.gouv.fr/fr/datasets/serie-hebdomadaire-par-espece-produits-de-la-mer/}{site data.gouv.fr}
}
\description{
Données hebdomadaires sur la pêche en France pour différents types de poissons
}
