% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Prenom}
\alias{Prenom}
\title{Prénoms des enfants nés en France}
\format{
Un tibble de 601461 observations contenant :
\itemize{
\item sexe : une variable catégorielle avec pour modalités \code{homme} et \code{femme},
\item prenom : le prénom considéré,
\item annee : l'année considérée,
\item nombre : le nombre d'enfants ayant reçu le prénom considéré pour l'année considérée
}
}
\source{
\href{https://www.insee.fr/fr/statistiques/2540004?sommaire=4767262}{site de l'INSEE}
}
\description{
Données annuelles indiquant, de 1900 à 2019, le nombre d'enfant
ayant reçu un prénom
}
