% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Queue}
\alias{Queue}
\title{File d'attente à la banque}
\format{
Un tibble contenant
\itemize{
\item id : l'identifiant de l'attente
\item time : l'heure d'arrivée
\item attente : la durée d'attente en minute
\item opération : la durée de l'opération au guichet
\item total : la temps total passé dans la banque
\item bank : identifiant de la banque (A, B ou C)
\item week : identitiant de la semaine (1, 2 ou 3)
\item day : identifiant du jour de la semaine (monday, tuesday, ...)
}
}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S2352340918315191#ec0006}{site de la revue Data in Brief}
}
\description{
Mesure de la fille d'attente dans des banques au Nigéria, 38875
observations, pour 3 banques pendant 3 semaines
}
\references{
\insertRef{NUNO:DEAL:NUNE:19}{statecoaes}
}
\keyword{datasets}
