% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{recut}
\alias{recut}
\title{Recode a classified variable in a tibble and return the tibble}
\usage{
recut(data, x, breaks = NULL)
}
\arguments{
\item{data}{a tibble}

\item{x}{the variable to recode}

\item{breaks}{a numerical vector of breaks}
}
\description{
\code{cut} take a numerical series as argument and return a class
according to a \code{break} vector ; this function recode a classified
series according to a vector of breaks which is a subset of the
original one
}
\author{
Yves Croissant
}
