% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RGP77}
\alias{RGP77}
\title{Données du recensement pour le département de la Seine et Marne (78)}
\format{
Un tibble de 101,987 observations contenant :
\itemize{
\item voitures : le nombre de voitures
\item pieces : le nombre de pièces du logement
\item enfants : le nombre d'enfants,
\item typemen : le type de famille : couple (\code{couple}) ou famille
mono-parentale (\code{monop}),
\item ER : le sexe du Rième enfant (de 1 à 11).
}
}
\source{
\href{https://www.insee.fr/fr/statistiques/4507685?sommaire=4508161}{site de l'INSEE}
}
\description{
Différentes informations concernant les familles en Seine et Marne,
issues du recensement de la population de 2017
}
