% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Salaires}
\alias{Salaires}
\title{Extraits de l'enquête DADS (Déclaration Annuelle des Données Sociales)}
\format{
Un tibble contenant :
\itemize{
\item secteur : secteur d'activité, \code{industrie},
\code{construction}, \code{commerce}, \code{services} et \code{administration},
\item age :  l'âge du salarié en année,
\item dept : code du département d'emploi,
\item heures : nombre d'heures annuelles travaillées,
\item sexe : sexe du salarié, \code{homme} ou \code{femme},
\item salaire : tranches de salaire, en euros par an,
\item sal_inf : limite inférieure de la tranche de salaire,
\item sal_sup : limite supérieure de la tranche de salaire
}
}
\source{
Enquête DADS 2015, site de l'\href{https://www.insee.fr/fr/statistiques/3536754}{INSEE}
}
\description{
L'enquête DADS indique de nombreuses caractéristiques des salariés,
en particulier le salaire annuel par tranche et le nombre d'heures
travaillées (les données sont l'extrait de la base correspondant à
la région Ile de France
}
