% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Seville}
\alias{Seville}
\title{Locations saisonnières à Seville}
\format{
Un tibble contenant :
\itemize{
\item price : le prix pour deux nuits
\item district : le disctrict dans lequel est situé la location
\item dindex : un index de richesse du district (1 pour le plus riche),
\item distance : distance aux centres d'intérêt majeurs de Seville
\item rating : moyenne des notes obtenues précédemment sur le site
\item beds : nombre de lits
\item surface : surface de l'appartement
\item parking : 1 si la location dispose d'un parking
\item season : \code{ls_wd} pour saison creuse - jour de semaine, \code{ls_we}
pour saison creuse - week end, \code{hs_wd} pour haute saison - jour de
semaine, \code{hs_we} pour haute saison - week end, \code{holy_week} pour la
semaine sainte et \code{april_fair}
}
}
\source{
Site de la revue \href{https://www.sciencedirect.com/science/article/pii/S2352340919310522#tbl5}{Data in Brief}
}
\description{
Prix et différentes caractéristiques de locations saisonnières dans
la ville de Séville, figurant sur le site bookings.com
}
\references{
\insertRef{SOLA:MARG:19}{statecoaes}
}
