% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Tabac}
\alias{Tabac}
\title{Consommation de tabac}
\format{
Un tibble de 32 observations contenant :
\itemize{
\item an : l'année d'obsrevation, de 2004 à 2019
\item bien : une variable catégorielle, \code{cigarette} ou \code{tabac} (paquet de tabac à rouler),
\item quant : millions de paquets vendus
\item prix : prix du paquet en euros
}
}
\source{
Tableau de bord tabac

Observatoire français des drogues et des toxicomanies

site de l'\href{https://www.ofdt.fr/statistiques-et-infographie/tableau-de-bord-tabac/}{ofdt}
}
\description{
Données annuelles concernant le prix unitaire et la quantité vendue
de cigarettes et de tabac à rouler
}
