% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Clap}
\alias{Clap}
\title{Nombre de pièces des logements}
\format{
Un tibble contenant :
\itemize{
\item secteur le secteur d'activité
\item taille la trance de taille de l'entreprise mesurée par le nombre de salariés,
\item nombre le nombre d'établissements,
\item salaries le nombre de salariés
}
}
\source{
\strong{Clap} (Connaissance locale de l'appareil productif), site de
l'\href{https://www.insee.fr/fr/statistiques/2021289}{INSEE}
}
\description{
Table de fréquence indiquant le nombre de pièce par logements
}
