% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Consommation}
\alias{Consommation}
\title{Consommation en France}
\format{
Un tibble contenant :
\itemize{
\item an : l'année d'observation
\item bien : un poste de consommation
\item depense : dépense de consommation en euros courants,
\item prix : indice de prix (base 100 en 2014),
Les postes de consommation sont :
\item alimentation,
\item alcool et tabac,
\item habillement,
\item logement,
\item ammeublement,
\item santé,
\item transport,
\item communications,
\item loisirs,
\item éducation,
\item restauration,
\item services divers,
\item solde territorial
}
}
\source{
Consomation des ménages en 2019, comptes nationaux annuels - base 2014

\href{https://www.insee.fr/fr/statistiques/4494154?sommaire=4494218}{site de l'INSEE}
}
\description{
Données annuelles de la consommation des ménages en France,
nomenclature par fonctions (alimentation, habillement, etc.)
}
