% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cont_table}
\alias{cont_table}
\title{Table de contingence}
\usage{
cont_table(
  data,
  x1,
  x2,
  abs = FALSE,
  pct = FALSE,
  marge = 0,
  pond = NULL,
  na.rm = TRUE,
  total = TRUE
)
}
\arguments{
\item{data}{un tibble}

\item{x1}{une première variable catégorielle}

\item{x2}{une seconde variable catégorielle}

\item{abs}{une valeur logique : \code{TRUE} pour renvoyer des fréquences
absolues, \code{FALSE} (la valeur par défaut) pour des fréquences
absolues,}

\item{pct}{une valeur logique : \code{TRUE} pour renvoyer des
pourcentages (20.2 au lieu de 0.202)}

\item{marge}{permet de calculer des fréquences conditionnelles en
lignes (\code{marge = 1}) ou en colonnes (\code{marge = 2}) ; avec la
valeur par défaut de \code{0}, les fréquences jointes et les
fréquences marginales sont calculées.}

\item{pond}{une éventuelle variable contenant les pondérations à
utiliser pour passer de l'échantillon à la population}

\item{na.rm}{la valeur par défaut est \code{TRUE}, les observations
pour lesquelles la valeur de \code{x} est manquante sont retirées de
l'échantillon}

\item{total}{si \code{TRUE} (valeur par défaut), un total est ajouté au
tableau}
}
\value{
un tibble
}
\description{
Une table de contingence rassemble les fréquences de toutes les
combinaisons de deux variables catégorielles sous la forme d'un
tableau à double entrée, avec les différentes modalités de la
première (seconde) variable en ligne (colonne)
}
\examples{

cont_table(Emploi, diplome, sexe)
cont_table(Emploi, diplome, sexe, pond = pondérations)
cont_table(Emploi, diplome, sexe, pond = pondérations, pct = TRUE)
cont_table(Emploi, diplome, sexe, pond = pondérations, pct = TRUE, marge = 1)
cont_table(Emploi, diplome, sexe, pond = pondérations, pct = TRUE, marge = 2)

}
\author{
Yves Croissant
}
