% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Emploi}
\alias{Emploi}
\title{Extraits de l'enquête emploi}
\format{
Un tibble contenant
\itemize{
\item activité : une variable catégorielle permettant de distinguer les
inactifs \verb{inactif, les chômeurs }chômeur\verb{et les actifs occupés}occupé`,
\item temps : temps de travail, \code{néant}, \code{partiel}, \code{complet} et \code{inconnu},
\item diplome : niveau de diplome,
\item age : l'âge en tranche,
\item sexe : le sexe (\code{F} ou \code{M}),
\item menage : type de ménage, une personne \code{unepers}, famille
monoparentale \code{monop}, couple sans enfants \code{cpl}, couple avec
enfants \code{cplenf} et autres ménages \code{autre},
\item ponderations : pondération pour chaque observation permettant de
passer de l'enquête à la population,
\item durchom : durée de l'épisode de chômage, en tranche,
\item durinact : durée de l'épisode d'inactivité, en tranche,
\item durentr : durée de travail dans l'entreprise, en tranche,
\item horaires : nombre d'heures de travail hebdomadaire, en tranche d'heures,
\item statut : variable catégorielle intégrant les deux variables \code{activité} et \code{temps}, soit \code{inactif}, \code{chômeur}, \code{partiel}, \code{complet} et \code{inconnu}
}
}
\source{
Enquête emploi 2018, site de l'\href{https://www.insee.fr/fr/statistiques/4191029}{INSEE}
}
\description{
L'enquête emploi donne de nombreuses informations sur les
caractéristiques des emplois occupés par un échantillon de personne
(actif/non-actif, chômage, temps partiel, niveau de diplôme, etc.
}
