% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{freq_table}
\alias{freq_table}
\alias{format.freq_table}
\title{Tables de Fréquence}
\usage{
freq_table(
  data,
  x,
  cols = "n",
  weights = NULL,
  na.rm = TRUE,
  total = TRUE,
  max = NA
)

\method{format}{freq_table}(x, ..., n = NULL, width = NULL, n_extra = NULL)
}
\arguments{
\item{data}{un tibble}

\item{x}{une variable catégorielle}

\item{cols}{une chaîne de caractère contenant les lettres \code{n} pour
nombre, \code{f} pour fréquence et \code{p} pour pourcentage ; les séries
cumulées sont obtenues en indiquant les mêmes lettres en
majuscule}

\item{weights}{une éventuelle variable contenant les pondérations à
utiliser pour passer de l'échantillon à la population}

\item{na.rm}{la valeur par défaut est \code{TRUE}, les observations
pour lesquelles la valeur de \code{x} est manquante sont retirées de
l'échantillon}

\item{total}{si \code{TRUE} (valeur par défaut), un total est ajouté au
tableau}

\item{max}{dans le cas où la variable est numérique entière, cet
argument indique que les valeurs supérieures ou égales à \code{max}
seront regroupées}

\item{...}{d'autres arguments (pour \code{format})}

\item{n}{le nombre de ligne à imprimer (pour \code{format})}

\item{width}{la largeur du tableau à imprimer (pour \code{format})}

\item{n_extra}{le nombre de colonnes suplémentaires décrites (pour \code{format})}
}
\value{
un objet de class \code{freq_table} qui hérite de \code{tbl_df}
}
\description{
Une table de fréquence est adaptée pour les variables
catégorielles, elle renvoit les différentes modalités et les
fréquences associées
}
\examples{

freq_table(Emploi, activité, "n")
freq_table(Emploi, activité, "nN")
freq_table(Emploi, activité, "fF", weights = pondérations)
freq_table(RGP78, enfants, "npNP")
freq_table(RGP78, enfants, "npNP", max = 5)
}
\author{
Yves Croissant
}
