% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_table.R
\name{hist_table.methods}
\alias{hist_table.methods}
\alias{mean.hist_table}
\alias{variance.hist_table}
\alias{stdev.hist_table}
\alias{madev.hist_table}
\alias{modval.hist_table}
\alias{quantile.hist_table}
\alias{median.hist_table}
\alias{medial.hist_table}
\alias{tantile}
\alias{gini}
\title{Methods for hist_table objects}
\usage{
\method{mean}{hist_table}(x, ..., na.rm = TRUE, tbl = FALSE)

\method{variance}{hist_table}(x, ..., na.rm = TRUE)

\method{stdev}{hist_table}(x, ..., na.rm = TRUE)

\method{madev}{hist_table}(x, center = c("median", "mean"), ..., na.rm = TRUE)

\method{modval}{hist_table}(x, ..., tbl = FALSE)

\method{quantile}{hist_table}(x, probs = c(0.25, 0.5, 0.75), tbl = FALSE, ...)

\method{median}{hist_table}(x, na.rm, ..., tbl = FALSE)

\method{medial}{hist_table}(x, tbl = FALSE)

tantile(x, probs = c(0.25, 0.5, 0.75), tbl = FALSE, ...)

gini(x)
}
\arguments{
\item{x}{a hist_table object,}

\item{...}{further arguments}

\item{na.rm}{a boolean, if \code{TRUE} missing values are removed}

\item{tbl}{if \code{TRUE} a tibble containing the class and the
densities is returned}

\item{probs}{the probabilities for which the quantiles have to be
computed}
}
\value{
a tibble if \code{tbl} is \code{TRUE}, a numeric otherwise
}
\description{
Functions and methods to compute the median, the mean, the mode,
the medial and quantiles for hist_table objects
}
\examples{

z <- Salaires \%>\% hist_table(salaire)
z \%>\% median
z \%>\% medial
z \%>\% modval
z \%>\% modval(tbl = TRUE)
z \%>\% quantile(probs = c(0.25, 0.5, 0.75))
z \%>\% quantile(probs = c(0.25, 0.5, 0.75), tbl = TRUE)
z \%>\% tantile(probs = c(0.25, 0.5, 0.75), tbl = TRUE)
}
\author{
Yves Croissant
}
