% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hist_table.methods}
\alias{hist_table.methods}
\alias{mean.hist_table}
\alias{modval}
\alias{modval.hist_table}
\alias{quantile.hist_table}
\alias{median.hist_table}
\alias{medial}
\alias{medial.hist_table}
\alias{tantile}
\title{Methods for hist_table objects}
\usage{
\method{mean}{hist_table}(x, ..., na.rm = TRUE, tbl = FALSE)

modval(x, ...)

\method{modval}{hist_table}(x, ..., tbl = FALSE)

\method{quantile}{hist_table}(x, probs = c(0.25, 0.5, 0.75), tbl = FALSE, ...)

\method{median}{hist_table}(x, tbl = FALSE)

medial(x, ...)

\method{medial}{hist_table}(x, tbl = FALSE)

tantile(x, probs = c(0.25, 0.5, 0.75), tbl = FALSE, ...)
}
\arguments{
\item{x}{a hist_table object,}

\item{...}{further arguments}

\item{na.rm}{a boolean, if \code{TRUE} missing values are removed}

\item{y}{the variable for which the quantiles have to be computed}

\item{dens}{if \code{TRUE} a tibble containing the class and the
densities is returned}

\item{prob}{the probabilities for which the quantiles have to be
computed}
}
\value{
a tibble if \code{dens} is \code{TRUE}, a numeric otherwise
}
\description{
Functions and methods to compute the median, the mean, the mode,
the medial and quantiles for hist_table objects
}
\examples{

z <- Salaires \%>\% hist_table(salaire, "dFM")
z \%>\% median
z \%>\% medial
z \%>\% modval
z \%>\% modval(dens = TRUE)
z \%>\% quantile(y = "F", prob = c(0.25, 0.5, 0.75))
z \%>\% quantile(y = "F", prob = c(0.25, 0.5, 0.75), dens = TRUE)
z \%>\% quantile(y = "M", prob = c(0.25, 0.5, 0.75), dens = TRUE)
}
\author{
Yves Croissant
}
