% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Ketchup}
\alias{Ketchup}
\title{Achat de Ketchup}
\format{
Un tibble contenant
\itemize{
\item chid: identifiant de l'achat
\item id : identifiant du ménage
\item week : identifiant de la semaine
\item choice : 1 si du ketchup a été acheté, 0 autrement
}
}
\source{
\href{https://qed.econ.queensu.ca/jae/}{Journal of Applied Econometrics data archive}
}
\description{
Comportement hebdomadaire d'achat de Ketchup par 3088 ménages au cours de l'année 1987
}
\references{
\insertRef{CHIN:ERDE:KEAN:09}{statecoaes}
}
\keyword{datasets}
