% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Vols}
\alias{Vols}
\title{Transport aérien en Europe}
\format{
Un tibble contenant :
\itemize{
\item date : la date d'observation (dernier jour du trimestre),
\item an : l'année d'observation,
\item trimestre : le trimestre d'observation
\item pays : code du pays (2 caractères)
\item passagers : nombre de passagers transportés
}
}
\source{
\href{https://ec.europa.eu/eurostat/fr/}{site d'Eurostat}
}
\description{
Ce jeux de données indique pour les différents pays de l'union
européenne le nombre de trajets en avion réalisés par trimestre
}
