% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Vols}
\alias{Vols}
\title{Transport aérien en Europe}
\format{
Un tibble contenant :
\itemize{
\item date : la date d'observation (premier jour du mois),
\item vols : nombe de vols
\item passagers : nombre de passagers transportés
}
}
\source{
Transport aérien de passagers par pays déclarant (avia_paoc)

site d'\href{https://ec.europa.eu/eurostat/fr/}{Eurostat}
}
\description{
Ce jeux de données indique pour l'Italie le nombre de vols et le
nombre de passagers par mois
}
